+++ arpeggio/run_pytest3_mypy_in_a_loop.sh
@@ -53,24 +53,24 @@ function pytest_loop {
             continue
         fi
 
-#        clr_green "*** MYPY Module strict **************************************************************************"
-#        python3 -m mypy "${my_dir}" --strict --no-warn-unused-ignores --follow-imports=skip
-#            # shellcheck disable=SC2181  # Check Exit Code directly
-#            if [[ "${?}" -gt 0 ]]; then
-#                clr_red "MYPY Error in Module"
-#                beep
-#                sleep "${sleeptime_on_error}"
-#                continue
-#            fi
+        clr_green "*** MYPY Module strict **************************************************************************"
+        python3 -m mypy "${my_dir}" --strict --no-warn-unused-ignores --follow-imports=skip
+            # shellcheck disable=SC2181  # Check Exit Code directly
+            if [[ "${?}" -gt 0 ]]; then
+                clr_red "MYPY Error in Module"
+                beep
+                sleep "${sleeptime_on_error}"
+                continue
+            fi
 
-#        clr_green "*** MYPY Imports strict *************************************************************************"
-#        python3 -m mypy "${my_dir}" --strict --no-warn-unused-ignores &>/dev/null
-#            # shellcheck disable=SC2181  # Check Exit Code directly
-#            if [[ "${?}" -gt 0 ]]; then
-#                clr_red "MYPY Error in Imports"
-#                sleep "${sleeptime_on_error}"
-#                continue
-#            fi
+        clr_green "*** MYPY Imports strict *************************************************************************"
+        python3 -m mypy "${my_dir}" --strict --no-warn-unused-ignores &>/dev/null
+            # shellcheck disable=SC2181  # Check Exit Code directly
+            if [[ "${?}" -gt 0 ]]; then
+                clr_red "MYPY Error in Imports"
+                sleep "${sleeptime_on_error}"
+                continue
+            fi
 
 
         sleep "${sleeptime_on_error}"
