+++ arpeggio/tests/regressions/issue_26/test_issue_26.py
@@ -1,6 +1,14 @@
-from ....cleanpeg import ParserPEG
+# proj
 
-def test_regex_with_empty_successful_match_in_repetition():
+try:
+    # for pytest
+    from ....cleanpeg import ParserPEG      # type: ignore # pragma: no cover
+except ImportError:                         # type: ignore # pragma: no cover
+    # for local Doctest
+    from cleanpeg import ParserPEG          # type: ignore # pragma: no cover
+
+
+def test_regex_with_empty_successful_match_in_repetition() -> None:
     grammar = \
             """
             rule = (subexpression)+
