+++ arpeggio/tests/regressions/issue_31/test_issue_31.py
@@ -1,10 +1,22 @@
-from .... import ParserPython, ZeroOrMore
+# proj
+try:
+    # imports for local pytest
+    from ....arpeggio import ParserPython   # type: ignore # pragma: no cover
+    from ....arpeggio import ZeroOrMore     # type: ignore # pragma: no cover
+except ImportError:                         # type: ignore # pragma: no cover
+    # imports for doctest
+    # noinspection PyUnresolvedReferences
+    from arpeggio import ParserPython       # type: ignore # pragma: no cover
+    from arpeggio import ZeroOrMore         # type: ignore # pragma: no cover
 
 
-def test_empty_nested_parse():
+def test_empty_nested_parse() -> None:
 
-    def grammar(): return [first]
-    def first(): return ZeroOrMore("second")
+    def grammar():
+        return [first]
+
+    def first():
+        return ZeroOrMore("second")
 
     parser = ParserPython(grammar)
