+++ arpeggio/tests/regressions/test_direct_rule_call.py
@@ -1,20 +1,32 @@
-import pytest
-from .... import SemanticAction, ParserPython
+# stdlib
+from typing import Any
+import pytest                                   # type: ignore
 
-def test_direct_rule_call():
-    '''
+# proj
+
+from ...arpeggio import ParserPython
+from ...peg_semantic_actions import SemanticAction
+
+
+def test_direct_rule_call() -> None:
+    """
     Test regression where in direct rule call semantic action is
     erroneously attached to both caller and callee.
-    '''
+    """
 
-    def grammar():  return rule1, rule2
-    def rule1():    return "a"
-    def rule2():    return rule1
+    def grammar() -> Any:
+        return rule1, rule2
+
+    def rule1() -> Any:
+        return "a"
+
+    def rule2() -> Any:
+        return rule1
 
     call_count = [0]
 
     class DummySemAction(SemanticAction):
-        def first_pass(self, parser, node, nodes):
+        def first_pass(self, parser, node, nodes) -> SemanticAction.first_pass:
             call_count[0] += 1
             return SemanticAction.first_pass(self, parser, node, nodes)
