+++ arpeggio/tests/test_error_reporting.py
@@ -6,10 +6,27 @@
 # License: MIT License
 #######################################################################
 
-import pytest
+import pytest   # type: ignore
 
-from .. import Optional, Not, ParserPython, NoMatch, EOF
-from .. import RegExMatch as _
+# proj
+try:
+    # imports for local pytest
+    from ..peg_expressions import Optional          # type: ignore # pragma: no cover
+    from ..peg_expressions import Not               # type: ignore # pragma: no cover
+    from ..peg_expressions import NoMatch           # type: ignore # pragma: no cover
+    from ..peg_expressions import EOF               # type: ignore # pragma: no cover
+    from ..peg_expressions import RegExMatch as _   # type: ignore # pragma: no cover
+    from ..arpeggio import ParserPython             # type: ignore # pragma: no cover
+
+except ImportError:                             # type: ignore # pragma: no cover
+    # imports for doctest
+    # noinspection PyUnresolvedReferences
+    from peg_expressions import Optional            # type: ignore # pragma: no cover
+    from peg_expressions import Not                 # type: ignore # pragma: no cover
+    from peg_expressions import NoMatch             # type: ignore # pragma: no cover
+    from peg_expressions import EOF                 # type: ignore # pragma: no cover
+    from peg_expressions import RegExMatch as _     # type: ignore # pragma: no cover
+    from arpeggio import ParserPython               # type: ignore # pragma: no cover
 
 
 def test_non_optional_precedence():
